Rename method and file name for routes
